# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:4000/apaman198/"
    },
    {
      "url": "https://localhost:4000/apaman198/"
    }
  ],
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0-oas3",
    "title": "Montaman API Document",
    "termsOfService": "https://monta.ir/apaman",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/cost/class": {
      "get": {
        "tags": [
          "cost",
          "new vsersion"
        ],
        "summary": "Get a list of costClass enums",
        "description": "",
        "operationId": "getCostClasses",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnumType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/cost/calculate": {
      "get": {
        "tags": [
          "cost",
          "new vsersion"
        ],
        "summary": "Get a list of costCalculate enums",
        "description": "",
        "operationId": "getCostCalculates",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnumType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/cost/occupation": {
      "get": {
        "tags": [
          "cost",
          "new vsersion"
        ],
        "summary": "Get a list of costOccupation enums",
        "description": "",
        "operationId": "getCostOccupation",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnumType"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EnumType": {
        "type": "object",
        "properties": {
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "PersianName": {
            "type": "string"
          }
        }
      },
...
    }
  }
}